Module content_blocker::repr

Structs

DomainMatcher

Request
A request that could be filtered.

Rule
A single rule, consisting of a condition to trigger this rule, and an action to take.

Trigger
A set of filters that determine if a given rule's action is performed.

Enums

Action
An action to take when a rule is triggered.

DomainConstraint
Conditions which restrict the set of matches for a particular trigger.

LoadType
The type of load that is being initiated.

Reaction
The action to take for the provided request.

ResourceType
The type of resource being requested.

ResourceTypeList
A potential list of resource types being requested.

Functions

process_rules_for_request_impl
Attempt to match the given request against the provided rules. Returns a list of actions to take in response; an empty list means that the request should continue unmodified.