Struct content_blocker::repr::Trigger

pub struct Trigger {
    pub url_filter: Regex,
    pub resource_type: ResourceTypeList,
    pub load_type: Option<LoadType>,
    pub domain_constraint: Option<DomainConstraint>,
}

A set of filters that determine if a given rule's action is performed.

Fields
url_filter: Regex

A simple regex that is matched against the characters in the destination resource's URL.

resource_type: ResourceTypeList

The classes of resources for which this trigger matches.

load_type: Option<LoadType>

The category of loads for which this trigger matches.

domain_constraint: Option<DomainConstraint>

Domains which modify the behaviour of this trigger, either specifically including or excluding from the matches based on string comparison.

Methods

impl Trigger

fn matches(&self, request: &Request) -> bool