Struct content_blocker::Request

pub struct Request<'a> {
    pub url: &'a Url,
    pub resource_type: ResourceType,
    pub load_type: LoadType,
}

A request that could be filtered.

Fields

url

The requested URL.

resource_type

The resource type for which this request was initiated.

load_type

The relationship of this request to the originating document.